Add Get api ById  in Session.

+++ src/main/java/com/school/session/SessionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SessionRepository extends PagingAndSortingRepository<Session, Long>{
+public interface SessionRepository extends PagingAndSortingRepository<Session,Integer>{
 
 }
